+++ rasa_core/training/interactive.py
@@ -985,14 +985,19 @@ def _validate_user_text(latest_message: Dict[Text, Any],
     This assumes the user message is a text message (so NOT `/greet`)."""
 
     parse_data = latest_message.get("parse_data", {})
-    entities = _as_md_message(parse_data)
+    text = _as_md_message(parse_data)
     intent = parse_data.get("intent", {}).get("name")
 
-    question = questionary.confirm(
-        "Is the NLU classification for '{}' with intent '{}' correct?"
-        "".format(entities, intent))
+    if intent is None:
+        print("The NLU classification for '{}' returned '{}'"
+              "".format(text, intent))
+        return False
+    else:
+        question = questionary.confirm(
+            "Is the NLU classification for '{}' with intent '{}' correct?"
+            "".format(text, intent))
 
-    return _ask_or_abort(question, sender_id, endpoint)
+        return _ask_or_abort(question, sender_id, endpoint)
 
 
 def _validate_nlu(intents: List[Text],